AtSpi2: Move State handling to AccessibleProvider.
